# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/system/hsm-config":{"put":{"tags":["system-configuration-controller"],"operationId":"setKnetConfiguration","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KNetConfigurationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}}},"security":[{"administrator":[]}]}},"/user":{"get":{"tags":["app-user-controller"],"operationId":"getUsers","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}}},"security":[{"administrator":[]}]},"post":{"tags":["app-user-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserRequest"}}},"required":true},"responses":{"201":{"description":"Created","headers":{"Location":{"description":"URI to the user created","style":"simple","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}}}}},"/system/refresh-key":{"post":{"tags":["system-configuration-controller"],"operationId":"refreshSystemKey","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"administrator":[]}]}},"/system/hsm-config/load":{"post":{"tags":["system-configuration-controller"],"operationId":"loadKnetConfiguration","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}}},"security":[{"administrator":[]}]}},"/system/admin-user":{"post":{"tags":["system-configuration-controller"],"operationId":"createSystemAdmin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserRequest"}}},"required":true},"responses":{"201":{"description":"Created","headers":{"Location":{"description":"URI to the user created","style":"simple","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}}}}},"/key":{"get":{"tags":["key-pair-controller"],"operationId":"getKeys","parameters":[{"name":"keyName","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/KeyResponse"}}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}}},"security":[{"user":[]}]},"post":{"tags":["key-pair-controller"],"operationId":"createKeyPair","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KeyPairGenerationRequest"}}},"required":true},"responses":{"201":{"description":"Created","headers":{"Location":{"description":"URI to the key created","style":"simple","schema":{"type":"string"}}},"content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}}},"security":[{"user":[]}]}},"/key/{keyUniqueIdentifier}/verify-signature":{"post":{"tags":["key-pair-controller"],"operationId":"verify","parameters":[{"name":"keyUniqueIdentifier","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignatureVerificationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignatureVerificationResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}}}}},"/key/{keyUniqueIdentifier}/sign":{"post":{"tags":["key-pair-controller"],"operationId":"sign","parameters":[{"name":"keyUniqueIdentifier","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignatureRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignatureResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}}},"security":[{"user":[]}]}},"/user/{username}":{"get":{"tags":["app-user-controller"],"operationId":"getUser","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}}},"security":[{"administrator":[]}]}},"/key/{keyUniqueIdentifier}":{"get":{"tags":["key-pair-controller"],"operationId":"getKey","parameters":[{"name":"keyUniqueIdentifier","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KeyResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}}},"security":[{"user":[]}]},"delete":{"tags":["key-pair-controller"],"operationId":"deleteKeyPair","parameters":[{"name":"keyUniqueIdentifier","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageResponse"}}}}},"security":[{"user":[]}]}},"/login":{"post":{"tags":["login-endpoint"],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"OK"},"403":{"description":"Forbidden"}}}}},"components":{"schemas":{"ErrorMessageResponse":{"type":"object","properties":{"error":{"type":"string"}}},"KNetConfigurationRequest":{"type":"object","properties":{"parameters":{"type":"object","additionalProperties":{"type":"string"}}}},"UserResponse":{"type":"object","properties":{"username":{"type":"string"},"authority":{"type":"string"}}},"RegisterUserRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"KeyPairGenerationRequest":{"type":"object","properties":{"keyAlgorithm":{"type":"string"},"keyParameter":{"type":"string"},"keyName":{"type":"string"}}},"SignatureVerificationResponse":{"type":"object","properties":{"validSignature":{"type":"boolean"}}},"SignatureVerificationRequest":{"type":"object","properties":{"base64EncodedData":{"type":"string"},"base64EncodedSignature":{"type":"string"},"signatureAlgorithm":{"type":"string"}}},"SignatureResponse":{"type":"object","properties":{"base64EncodedSignature":{"type":"string"},"keyPairUniqueIdentifier":{"type":"string"},"base64EncodedPublicKey":{"type":"string"}}},"SignatureRequest":{"type":"object","properties":{"hashingAlgorithm":{"type":"string"},"base64EncodedData":{"type":"string"}}},"KeyResponse":{"type":"object","properties":{"keyPairUniqueIdentifier":{"type":"string"},"keyAlgorithm":{"type":"string"},"publicKey":{"type":"string"},"keyName":{"type":"string"}}}},"securitySchemes":{"administrator":{"type":"http","description":"Schema referring to a JWT obtained by an administrator user.","scheme":"bearer","bearerFormat":"JWT"},"user":{"type":"http","description":"Schema referring to a JWT obtained by a non-administrator user.","scheme":"bearer","bearerFormat":"JWT"}}}}